+++ Sources/Core/Client/API/Client+RxEvents.swift
@@ -53,9 +53,9 @@ public extension Reactive where Base == Client {
         let events: Observable<WebSocket.Response>
         
         if let channel = channel {
-            events = channel.query(options: .watch).flatMapLatest { _ in Client.shared.webSocket.response }
+            events = channel.query(options: .watch).flatMapLatest { [unowned base] _ in base.webSocket.rx.response }
         } else {
-            events = base.webSocket.response
+            events = base.webSocket.rx.response
         }
         
         return connection.connected()
